Velocity/PromptEnhancement/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
arxiv==2.1.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
black==24.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.5.14
Deprecated==1.2.18
distro==1.9.0
duckduckgo_search==8.0.1
durationpy==0.9
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.11.0
fastapi==0.115.12
feedparser==6.0.10
filelock==3.18.0
flake8==7.0.0
Flask==3.0.2
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
# google-generativeai==0.8.4  # Deprecated - use langchain-google-genai instead
googleapis-common-protos==1.70.0
groq==0.22.0
grpcio==1.71.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.24
langchain-community==0.3.16
langchain-core==0.3.66
langchain-google-community==2.0.7
langchain-google-genai==2.1.5
langchain-groq==0.3.2
langchain-nvidia-ai-endpoints==0.3.9
langchain-openai==0.3.14
langchain-postgres==0.0.14
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.4.0
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.64
langid==1.1.6
langsmith==0.3.45
limits==5.2.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.34.1
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.76.2
openapi-schema-pydantic==1.2.4
openpyxl==3.1.5
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.10.16
ormsgpack==1.9.1
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.4
pgvector==0.3.6
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
posthog==4.0.1
primp==0.15.0
propcache==0.3.1
proto-plus==1.22.3
protobuf==4.23.4
psycopg==3.2.9
psycopg-pool==3.2.6
psutil==6.1.0
pyarrow==19.0.1
aiofiles==24.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.2
pyparsing==3.2.3
pypdf==5.4.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
schedule==1.2.2
scikit-learn==1.3.0
scipy==1.15.2
seaborn==0.13.2
serpapi==0.1.5
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
streamlit==1.44.1
sympy==1.14.0
tavily-python==0.7.2
tenacity==8.5.0
textblob==0.15.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0